Work around for P/Invoke AccessViolationException
The issue is on ARM builds for Windows UWP. For the C# P/Invoke API wrapper, any native method that returns the YogaValue struct throws the AccessViolationException. The issue is not with structs in general, as returning the YogaSize / YGSize struct works fine. The issue seems to be limited to structs that have an enum member.

I tried a number of things to resolve the issue without changing the underlying native API for Windows. I read the ARM documentation and saw reference to variable enum sizes based on the number of enum members, so I tried to use a number of different UnmanagedType values in a [MarsalAs()] attribute on the enum member of the C# struct declaration.

What ultimately worked was to return a pointer to the location of the struct, and use the System.Runtime.InteropServices.PtrToStructure API to read the struct data from that pointer. I added a few new macros that will return the struct address on Windows only, other builds are not affected.

Note, I have not tested the impact of this change on Xamarin.iOS, Xamarin.Mac, or Xamarin.Android, and I suspect some changes will need to be made to accomodate these target frameworks (cc @rmarinho).

The downside to this PR is that System.Runtime.InteropServices.PtrToStructure returns System.Object, which means the YogaValue struct will incur boxing, ultimately having an impact on the heap and Gen0 GC collections. However, I think this is still a reasonable workaround until we can figure out why the default struct marshalling behavior for P/Invoke is throwing the AVE for structs with enums.

Fixes #415