Add basic API for cloning a root Node
This is the first step to a persistent variant of Yoga.
This allows the tree to get a new layout without changing the first tree.

If, as result of layout, more nodes needs to fork, we clone those before
touching them. In practice, in the current algorithm, that means that we
clone all children of a Node if need to perform layout on any of them.