Add short cut to remove all children
When we have a fresh clone, we typically want to replace all the children
with child clones. An easy way to do that is to simply drop the child set.
This adds a fast path for deleting all children of a node, that by passes
the need to first clone all children and then remove them one by one.