YogaKit: flexBasis can now be set to Auto
We must be able to set flexBasis to Auto through YogaKit. Previously doing that we causing an assert in YG_AUTO_VALUE_PROPERTY.
flexBasis set to Auto means the length is equal to the length of the item. If the item has no length specified, the length will be according to its content.