set default value of 0 when NaN returned from layout
Occasionally YogaKit is returning NaN values after doing layout on a view.  On iOS, this causes an exception to occur when NaN is passed as the value of the frame size or origin.  This sets a default value of 0 rather than allowing NaN to be set for the view frame size or origin.