Add explicit defaulted copy ctor to `Values`
Summary:
`Values` is a wrapper to story an array of YGValue's as CompactValues.

From https://github.com/facebook/yoga/issues/1174 we see a warning `Wdeprecated-copy` beacuse a user-defined copy constructor is not present, but a user-defined asignment operator is (the defaulted one). This adds an explicitly defaulted copy contructor which should silence the warning I think.

Changelog:
[Internal]

Differential Revision: D41447490

fbshipit-source-id: 1c1911b46aec941e481484e814dae43f56312f17