Undeprecate Config setting APIs
Summary:
YGNode::setConfig was deprecated as part of D15416474 in an apparent goal to remove config pointers per-Node.

While I don't know the history of the motivation here, these config pointers were never removed, and we will be doubling down on per-node configs for StrictLayout, so we will want to undeprecate this.

This also exposes functions to the public C ABI, but I didn't spend the effort to create language projections for it.

Differential Revision: D45133646

fbshipit-source-id: 15793611d8a665e8d8ecc9e38663a7f2ebd6adce