Undeprecate Config setting APIs (#36990)
Summary:
X-link: https://github.com/facebook/react-native/pull/36990

YGNode::setConfig was deprecated as part of D15416474 in an apparent goal to remove config pointers per-Node.

While I don't know the history of the motivation here, these config pointers were never removed, and we will be doubling down on per-node configs for StrictLayout, so we will want to undeprecate this.

This also exposes functions to the public C ABI, but I didn't spend the effort to create language projections for it.

Differential Revision: D45133646

fbshipit-source-id: eb148802732e76b5e2e7bb85e3a3e0da229df1e4