Move YogaKit to YGErrataClassic
Summary:
YogaKit integrates Yoga with UIKit as a higher level framework. It does not expose config setting to users.

We set YGErrataClassic for now to prioritize compatibility instead of conformance (YogaKit is relatively used in fbsource as well).

I'm also tempted to remove the usage of ExperimentalWebFlexBasis since last I heard rozelle thought it was generally broken, but I am a bit afraid to if it has been enabled so long, and is used in many cases in Meta.

Differential Revision: https://internalfb.com/D45298803

fbshipit-source-id: 266efce67de5bf9fb24982267ee4f4df59151a9f