Move generated Jest tests to TypeScript
Summary:
Outputs tests as TypeScript, along with using/testing the new form of enums imported directly from the package.

We need to change how we are telling Jest which variant to run, so that tests can import enums from "yoga-layout" and have it resolve to the entrypoint which has a binary which has already been built.

Differential Revision: D45723545

fbshipit-source-id: 300accf064a9efaf145204a1023803b003758cb7