Mark YogaKit Podspec as Deprecated
Summary:
We are deprecating YogaKit as part of Yoga 2.0, with that version planned as the last release. Mark the podspec as deprecated.

We are also deprecating the Android ViewGroup in a similar way but from the googling I did there is not an equivalent way to do that for Maven artifacts.

Differential Revision: D46663201

fbshipit-source-id: 0f5c892589f76855adb3136f28d2128e9551a2d1