Docusarus: Init Project
Summary:
This starts us transitioning the Yoga website off of an ancient version of Gatsby by initializing a new Docusaurus project. This commit adds a new stock project, to be filled in later, then fixes license lint errors.

Note that this seems to have init'd using npm, but in the next diff I convert this to yarn workspaces anyway.

Differential Revision: D46884434

fbshipit-source-id: 26e2ecba057ae99c4ef878e4cd21c8d21a25c2b5