Docusarus: Init Project
Summary:
This starts us transitioning the Yoga website off of an ancient version of Gatsby by initializing a new Docusaurus project. This commit adds a new stock project, to be filled in later, then fixes license lint errors.

Note that this seems to have init'd using npm, but in the next diff I convert this to yarn workspaces anyway.

Differential Revision: D46884434

fbshipit-source-id: 705c12b44e0cc9f601ac6d5b6d64c945bc2d1d7c