Docusaurus: Use yarn workspaces
Summary:
Use yarn workspaces to consolidate the lockfile, and to let the website use our local Yoga package. This was not possible with the old version of Gatsby we were previously using.

This does not yet sync the versions of different packages, or move config files to the root of the repo. That will happen later after removing the original website package.

Differential Revision: D46884433

fbshipit-source-id: ac2793e6ba6251c9c89748b6c0906a23d0d0127c