Fixup Yoga UT compilation on non-clang compilers
Summary:
Yoga is compiled on more than clang, both internally, and in OSS. D47345669 added unguarded usage of `
#pragma clang diagnostic push` which will cause unrecognized pragma errors on other platforms.

The code here is flagged for implicitly converting enum to float. This is test code which is using the enum as an arbitrary differentiating input, so I just changed this to an explicit cast.

Differential Revision: D47460301

fbshipit-source-id: 237b7eef459e87c0dd094e4297dcd280d0113388